# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Private Practice Management & Client Portal API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /calendar:
    get:
      operationId: listCalendarEntries
      tags: [calendar]
      summary: Calendar updates visible to the caller
      description: >
        Scoped twice: to the caller's workspace, and to entries whose audience
        includes their role or their user id. A client never receives a
        staff-only notice.
      responses:
        "200":
          description: Calendar entries
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/CalendarEntry"
        "403":
          description: Forbidden
    post:
      operationId: createCalendarEntry
      tags: [calendar]
      summary: Post a calendar update (Admin and Senior Advocate only)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CalendarEntryInput"
      responses:
        "201":
          description: Entry created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CalendarEntry"
        "403":
          description: Forbidden
  /calendar/{id}:
    patch:
      operationId: updateCalendarEntry
      tags: [calendar]
      summary: Move or edit a calendar entry (Admin and Senior Advocate only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CalendarEntryUpdate"
      responses:
        "200":
          description: Entry updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CalendarEntry"
        "403":
          description: Forbidden
        "404":
          description: Not found in this workspace
    delete:
      operationId: deleteCalendarEntry
      tags: [calendar]
      summary: Remove a calendar update (Admin and Senior Advocate only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Removed
        "403":
          description: Forbidden
        "404":
          description: Not found in this workspace
...
components:
  schemas:
...
    CalendarEntry:
      type: object
      required: [id, workspaceId, title, kind, entryDate, audience, createdAt]
      properties:
        id:
          type: integer
        workspaceId:
          type: integer
        title:
          type: string
        notes:
          type: ["string", "null"]
        kind:
          type: string
          enum: [hearing, filing, meeting, note]
        entryDate:
          type: string
        entryTime:
          type: ["string", "null"]
        caseId:
          type: ["integer", "null"]
        caseTitle:
          type: ["string", "null"]
        audience:
          type: string
        audienceLabel:
          type: ["string", "null"]
        createdBy:
          type: ["string", "null"]
        createdByRole:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
    CalendarEntryInput:
      type: object
      required: [title, entryDate]
      properties:
        title:
          type: string
          minLength: 1
        notes:
          type: string
        kind:
          type: string
          enum: [hearing, filing, meeting, note]
          default: note
        entryDate:
          type: string
          description: YYYY-MM-DD
        entryTime:
          type: string
        caseId:
          type: integer
        audience:
          type: string
          default: all
          description: "all | staff | role:<role> | user:<clerkId>"
    CalendarEntryUpdate:
      type: object
      properties:
        title:
          type: string
          minLength: 1
        notes:
          type: string
        kind:
          type: string
          enum: [hearing, filing, meeting, note]
        entryDate:
          type: string
        entryTime:
          type: string
        caseId:
          type: integer
        audience:
          type: string
...